pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homeassistant-mcp"
version = "0.3.9"
description = "Home Assistant MCP integration for Lovelace dashboard authoring"
requires-python = ">=3.13"

[project.optional-dependencies]
test = [
  "mcp==1.26.0",
  "pytest==9.0.0",
  "pytest-homeassistant-custom-component==0.13.316",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra"

[tool.setuptools.packages.find]
include = ["custom_components*"]
exclude = ["docs*", "specs*", "tests*", "lru*"]

[tool.setuptools.package-data]
"custom_components.homeassistant_mcp" = ["*.json", "translations/*.json", "brand/*.png"]

[tool.unittest]
start-directory = "tests"
pattern = "test_*.py"
